Small updates to get prune_ratio tree induction working; cat(s,a,ns) if no features

+++ rewards/graph.py
@@ -1,7 +1,7 @@
 import networkx as nx
 from numpy import array
 from numpy.random import default_rng
-from torch import tensor, isnan, zeros, device, cuda
+from torch import tensor, float32, isnan, zeros, device, cuda
 import matplotlib.pyplot as plt
 
 
@@ -41,9 +41,9 @@ class PreferenceGraph:
         return matrix
 
     def tensorise(self, s_a_ns_list):
-        return tensor(array([s  for s,_,__ in s_a_ns_list]), device=self.device), \
-               tensor(array([a  for _,a,__ in s_a_ns_list]), device=self.device), \
-               tensor(array([ns for _,_,ns in s_a_ns_list]), device=self.device)
+        return tensor(array([s  for s,_,__ in s_a_ns_list]), dtype=float32, device=self.device), \
+               tensor(array([a  for _,a,__ in s_a_ns_list]), dtype=float32, device=self.device), \
+               tensor(array([ns for _,_,ns in s_a_ns_list]), dtype=float32, device=self.device)
 
     def add_episode(self, states, actions, next_states, **kwargs):
         # TODO: Storing both states and next_states is wasteful! Just store T+1 states alongside T actions
